Delete Quiz and Category using SweetAlert popup module

+++ Frontend/onlineexamportal-ui/src/app/components/admin/view-category/view-category.component.ts
@@ -17,11 +17,11 @@ export class ViewCategoryComponent implements OnInit {
   constructor(private categoryService: CategoryService, private sharedCategoryData: CategoryDataService) { }
  
   ngOnInit(): void {
-    this.getAllCategories();
+    this.getAllCategoriesOrderByIdDesc();
   }
 
-  getAllCategories(): void {
-    this.categoryService.getAllCategories().subscribe((response: any) => {
+  getAllCategoriesOrderByIdDesc(): void {
+    this.categoryService.getAllCategoriesOrderByIdDesc().subscribe((response: any) => {
       this.categories = response;
     },
     (error)=> {
@@ -34,22 +34,32 @@ export class ViewCategoryComponent implements OnInit {
     })
   }
 
-  deleteCategory(id: any) {
-    this.categoryService.deleteCategoryById(id).subscribe((response: any) => {      
-      Swal.fire({
-        title: 'Success!!',
-        text: 'Category has been deleted successfully...',
-        icon: 'success',
-        confirmButtonText: 'OK'
-      })
-      this.getAllCategories();
-    },(error)=> {
-      Swal.fire({
-        title: 'Error!',
-        text: 'Error in deleting category.',
-        icon: 'error',
-        confirmButtonText: 'OK'
-      })
+  deleteCategory(id: any, title: any) {
+    Swal.fire({
+      title: title,
+      text: "Are you sure to delete this category?",
+      icon: "warning",
+      confirmButtonText: "Delete",
+      showCancelButton: true
+    }).then((response) => {
+      if(response.isConfirmed) {
+        this.categoryService.deleteCategoryById(id).subscribe((response: any) => {      
+          Swal.fire({
+            title: 'Success!!',
+            text: 'Category has been deleted successfully...',
+            icon: 'success',
+            confirmButtonText: 'OK'
+          })
+          this.getAllCategoriesOrderByIdDesc();
+        },(error)=> {
+          Swal.fire({
+            title: 'Error!',
+            text: 'Error in deleting category.',
+            icon: 'error',
+            confirmButtonText: 'OK'
+          })
+        })
+      }
     })
   }
